weather/pyproject.toml
[project]
name = "weather-mcp"
version = "0.1.0"
description = "Weather MCP server using NWS API"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]",
    "httpx",
    "python-dotenv",
    "google-adk",
    "ag-ui-adk",
    "fastapi",
    "uvicorn",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

